refactor: Active Session을 위한 패키지 분리
