+++ src/main/java/com/waruru/areyouhere/active/ActiveSessionService.java
@@ -1,0 +1,4 @@
+package com.waruru.areyouhere.active;
+
+public interface ActiveSessionService {
+}
