+++ src/main/java/com/waruru/areyouhere/active/ActiveSessionServiceImpl.java
@@ -1,0 +1,9 @@
+package com.waruru.areyouhere.active;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class ActiveSessionServiceImpl {
+}
