--- src/main/java/com/waruru/areyouhere/session/domain/repository/AuthCodeRedisRepository.java
@@ -1,9 +1,0 @@
-package com.waruru.areyouhere.session.domain.repository;
-
-import com.waruru.areyouhere.attendance.domain.entity.CurrentSessionAttendanceInfo;
-import java.util.Optional;
-import org.springframework.data.repository.CrudRepository;
-
-public interface AuthCodeRedisRepository extends CrudRepository<CurrentSessionAttendanceInfo, String>{
-
-}
